Colocate ProfileWithRelationship with its only consumer
The component was only used in friend-search-overlay.tsx; inline it
there and remove the standalone file.
