--- src/components/profile-with-relationship.tsx
@@ -1,105 +1,0 @@
-import { Check, Send, ThumbsUp, UserCheck, X } from 'lucide-react'
-import { Button } from '@/components/ui/button'
-import { AvatarIconRow } from '@/components/ui/avatar-icon'
-import { ConfirmDestructiveActionDialog } from '@/components/confirm-destructive-action-dialog'
-import { useFriendRequestAction } from '@/features/social/hooks'
-import { IconSizedLoader } from '@/components/ui/loader'
-import { uuid } from '@/types/main'
-import { useOnePublicProfile } from '@/features/social/public-profile'
-
-export function ProfileWithRelationship({ uid }: { uid: uuid }) {
-	const { data: profile } = useOnePublicProfile(uid)
-	const inviteResponseMutation = useFriendRequestAction(uid)
-	const isMostRecentByThem = profile?.relation?.most_recent_uid_by === uid
-
-	return !profile ? null : (
-		<AvatarIconRow {...profile}>
-			<div className="flex flex-row gap-2">
-				{inviteResponseMutation.isPending ? (
-					<IconSizedLoader />
-				) : inviteResponseMutation.isSuccess ? (
-					<span className="rounded-squircle size-8 rounded-full bg-green-600 p-1">
-						<Check className="size-6 text-white" />
-					</span>
-				) : !profile.relation || profile.relation.status === 'unconnected' ? (
-					<Button
-						variant="default"
-						className="size-8"
-						size="icon"
-						aria-label="Send friend request"
-						onClick={() => inviteResponseMutation.mutate('invite')}
-					>
-						<Send className="mt-[0.1rem] mr-[0.1rem] size-6" />
-					</Button>
-				) : profile.relation.status === 'pending' && isMostRecentByThem ? (
-					<>
-						<Button
-							variant="default"
-							className="size-8"
-							size="icon"
-							aria-label="Accept pending invitation"
-							onClick={() => inviteResponseMutation.mutate('accept')}
-						>
-							<ThumbsUp />
-						</Button>
-						<ConfirmDestructiveActionDialog
-							title="Decline this invitation"
-							description="Please confirm whether you'd like to decline this invitation"
-						>
-							<Button
-								variant="neutral"
-								className="size-8"
-								size="icon"
-								aria-label="Decline pending invitation"
-							>
-								<X className="size-6 p-0" />
-							</Button>
-							<Button
-								variant="red"
-								aria-label="Confirm: Decline friend request"
-								onClick={() => inviteResponseMutation.mutate('decline')}
-							>
-								{inviteResponseMutation.isPending ? (
-									<IconSizedLoader />
-								) : inviteResponseMutation.isSuccess ? (
-									<Check className="size-6 text-white" />
-								) : (
-									<>Confirm</>
-								)}
-							</Button>
-						</ConfirmDestructiveActionDialog>
-					</>
-				) : profile.relation?.status === 'pending' && !isMostRecentByThem ? (
-					<ConfirmDestructiveActionDialog
-						title={`Cancel this request`}
-						description={`Please confirm whether you'd like to cancel this friend request`}
-					>
-						<Button
-							variant="neutral"
-							className="size-8"
-							size="icon"
-							aria-label="Cancel friend request"
-						>
-							<X className="size-6 p-0" />
-						</Button>
-						<Button
-							variant="red"
-							aria-label="Confirm: Cancel friend request"
-							onClick={() => inviteResponseMutation.mutate('cancel')}
-						>
-							{inviteResponseMutation.isPending ? (
-								<IconSizedLoader />
-							) : inviteResponseMutation.isSuccess ? (
-								<Check className="size-6 text-white" />
-							) : (
-								<>Confirm</>
-							)}
-						</Button>
-					</ConfirmDestructiveActionDialog>
-				) : profile.relation.status === 'friends' ? (
-					<UserCheck className="size-6 p-0" />
-				) : null}
-			</div>
-		</AvatarIconRow>
-	)
-}
